requirements.txt
streamlit>=1.31.0
pdfplumber
groq
google-genai
python-dotenv
reportlab
plotly
pandas
mysql-connector-python
bcrypt
